requirements_dev.txt
-r requirements.txt
ipython
pylint
pylint-django
coverage
model-mommy
django-coverage
django-debug-toolbar
selenium
